Remove dependency on activesupport
Fudge uses activesupport in only a handful of places, and purely for
the convenience of the fudge developers.  It seems unreasonable to
require users of fudge to install this dependency for our convenience.
It can also cause dependency problems where a fudge user locks
activesupport to a specific version but a later version is installed.

This change removes the dependency, and provides our own
implementations for the activesupport features we were using.